requirements.txt
pandas
numpy
seaborn
matplotlib
rembg
notebook
pyYAML
tqdm
ensure==1.0.2
joblib
Flask
Flask-Cors
types-PyYAML
